Fix a problem with security relevant file rights
The file /etc/nullmailer/remotes is a file where also password could be
configured. Therefore the original file is owned by mail and only
readable by user mail.

If there is a global setting for owner and mode as defaults for file as
it is very common, this would leak the data to the world. Setting that
explicit for the file fixes this issue.